split productgateway from quotegateway
