+++ API_Gateway/API_Gateway/Controllers/API_GatewayController.cs
@@ -33,49 +33,4 @@ namespace API_Gateway.Controllers
             throw new NotImplementedException();
         }
     }
-
-
-    [Route("api-crm/")]
-    [ApiController]
-    public class API_ProductGatewayController : ControllerBase
-    {
-        private readonly IProductBackingService _productDB;
-        public API_ProductGatewayController(IProductBackingService productBS)
-        {
-            _productDB = productBS;
-        }
-
-        [HttpGet]
-        [Route("product")]
-        public List<ProductBsDTO> GetAll()
-        {
-            return _productDB.GetAllProduct().Result;
-        }
-
-        [HttpPost]
-        [Route("product")]
-        public ProductBsDTO PostProduct([FromBody] ProductBsDTO newproductDTO)
-        {
-            return _productDB.AddNew(newproductDTO).Result;
-        }
-
-        [HttpPut]
-        [Route("product/{id}")]
-        public ProductBsDTO PutProduct(string id, [FromBody] ProductBsDTO updateProduct)
-        {
-
-
-            return _productDB.Update(updateProduct, id).Result;
-
-        }
-
-        [HttpDelete]
-        [Route("{id}")]
-        public bool Delete(string id)
-        {
-            return _productDB.Delete(id).Result;
-        }
-    }
-
-
 }
